Optimize number data types
Simple change, certain fields of entities and composite key classes did not need to be 32-bit integers and could work with smaller data types.

I looked at the MySQL database table and column structure, then changed some Java data types accordingly.

int = int (4 bytes)
smallint = short (2 bytes)
tinyint = byte (1 byte)

+++ src/main/java/com/carpi/carpibackend/dto/CourseDto.java
@@ -15,7 +15,7 @@ public class CourseDto {
     private String department;
 
     @Min(0)
-    private int code;
+    private short code;
 
     @NotNull
     private String title;
@@ -24,10 +24,10 @@ public class CourseDto {
     private String description;
 
     @Min(0)
-    private short creditMin;
+    private byte creditMin;
 
     @Min(0)
-    private short creditMax;
+    private byte creditMax;
 
     @NotNull
     private String[] semesterList;
